requirements.txt
python-telegram-bot==21.6
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.48.0
